FloatingPoint: Clean up and document header, format.
This cleans up the src/util/floatingpoint.h.in header to conform to the
code style guidelines of CVC4. This further attempts to fully document
the header. The main changes (except for added documentation) are renames
of member variables, and a more explicit naming of CVC4 wrapper types vs
symFPU trait types. Else, it's mainly reordering and formatting.