[proof-new] Make prop engine proof producing
This makes the prop engine use the ProofCnfStream rather than the regular
CnfStream when proofs are enabled, as well as use the prop engine proof manager
to produce and check prop engine proofs when requested.

In a next PR Minisat will be made producing, thus completing proof production
for the prop engine.