Learn equalities involving Boolean variables
Previously, the circuit propagator was not learning literals of the form `(= x
t)` where `x` is Boolean, since this term was not treated as a theory literal.
This commit changes that, which improves performance significantly, since it
allows the elimination of Boolean variables, which, in turn, can make the
justification heuristic much more effective.

Signed-off-by: Andres Noetzli <noetzli@amazon.com>